requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
gunicorn>=20.1.0
beautifulsoup4>=4.9.0
email-validator>=1.1.0
python-dotenv>=0.19.0
requests>=2.25.0
trafilatura>=1.4.0
psycopg2-binary>=2.9.0
numpy>=1.20.0
scikit-learn>=1.0.0
joblib>=1.1.0
